move `ClosureRegionRequirements` to `rustc_borrowck`
